include of these chances: For the "Next Data Update", we'll need to have a countdown going down for each hour and a minute (00:01, 01:01, 02:01, etc.), so it should be HH : MM : SS.
Also, I think the "X" values on the last 24h line graph are not correct. It seems like they are opposite
I want it to go from Left to Right
with Right being the newest and left being the oldest

*The final thing is we're going to have to change the times from UTC to EST. So we will have to convert it on the site's end.

So this is actually done every 6 hours. So just make it go to the 1st minute of every 6 hours (00:01:00, 06:01:00, etc.), like how the next data update is.

On the filter button, can you please add an icon for the dropdown? Like an upside down triangle to indicate it is a dropdown menu? Possibly have it change depending if the menu is opened or closed, if closed, upside down triangle, if open, upside triangle.

Yeah it auto changes depending on your system settings.

https://gyazo.com/fc4dfd84e29046fcef14897d0a2800f0
You can see how it works here

A working Google search clone made with Next.js & Tailwind CSS. - GitHub - jakeva/google-search-clone: A working Google search clone made with Next.js & Tailwind CSS.

We are also going to need a dark mode. It should auto change depending on the system's preferences, like my other site:
https://google.jakevalenzuela.com/
Source code is here (if you need to look at how I do it):
https://github.com/jakeva/google-search-clone
GitHub - jakeva/google-search-clone: A working Google search clone made with Next.js & Tailwind CSS.

For the "Next Prediction Update", feel free for now to just put it at "N/A" since I'm not too sure how often we'll do this

And here the "Available Spaces" should be "Garage X Available Spaces" so like "Garage A Available Spaces"